KEPCO_fall_2024/Week1/Multimodal LLM/requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aioprocessing==2.0.1
aiosignal==1.3.1
analytics-python==1.2.9
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backoff==2.2.1
bitsandbytes==0.44.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
debugpy==1.8.6
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
emoji==2.14.0
evaluate==0.4.3
executing==2.1.0
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gql==3.5.0
gradio==4.44.1
gradio_client==1.3.0
graphql-core==3.2.4
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
hf_transfer==0.1.8
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
janus==1.0.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.51.0
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft==0.13.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.1
ruff==0.6.8
safetensors==0.4.5
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.15.0
setproctitle==1.3.3
setuptools==75.1.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-cpu==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
tiktoken==0.8.0
timm==1.0.9
tokenizers==0.20.0
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
triton==3.0.0
trl==0.11.1
typer==0.12.5
typing_extensions==4.12.2
tyro==0.8.11
tzdata==2024.2
unsloth==2024.9.post4
urllib3==2.2.3
uuid_utils==0.9.0
uvicorn==0.31.0
wandb==0.18.3
wcwidth==0.2.13
weave==0.51.12
websockets==12.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xformers==0.0.28.post1
xxhash==3.5.0
yarl==1.13.1
